[Refactor] mod useEditProfileModal #775
