+++ hooks/modal/useEditProfileModal.ts
@@ -1,80 +1,27 @@
-import React, { useEffect, useState, Dispatch, SetStateAction } from 'react';
-import { useRecoilState, useSetRecoilState } from 'recoil';
+import { SetterOrUpdater, useSetRecoilState } from 'recoil';
+import { ProfileBasic } from 'types/userTypes';
 import instance from 'utils/axios';
 import { errorState } from 'utils/recoil/error';
 import { modalState } from 'utils/recoil/modal';
-import { profileState } from 'utils/recoil/user';
+import { EditedProfile } from 'components/modal/profile/EditProfileModal';
 
-interface EditedProfile {
-  racketType: string;
-  statusMessage: string;
-  snsNotiOpt: 'NONE' | 'SLACK' | 'EMAIL' | 'BOTH';
-}
-
-const CHAR_LIMIT = 30;
-
-type useEditProfileModalResult = [
+type useEditProfileModalProps = [
   boolean,
   boolean,
-  Dispatch<SetStateAction<boolean>>,
   EditedProfile,
-  ({
-    target: { name, value },
-  }:
-    | React.ChangeEvent<HTMLTextAreaElement>
-    | React.ChangeEvent<HTMLInputElement>) => void,
-  () => Promise<void>,
-  () => void
+  SetterOrUpdater<ProfileBasic>
 ];
 
-const useEditProfileModal = (): useEditProfileModalResult => {
+type useEditProfileModalResult = [() => Promise<void>, () => void];
+
+const useEditProfileModal = ([
+  slack,
+  email,
+  editedProfile,
+  setProfile,
+]: useEditProfileModalProps): useEditProfileModalResult => {
   const setError = useSetRecoilState(errorState);
   const setModal = useSetRecoilState(modalState);
-  const [profile, setProfile] = useRecoilState(profileState);
-  const [editedProfile, setEditedProfile] = useState<EditedProfile>({
-    racketType: profile.racketType,
-    statusMessage: '',
-    snsNotiOpt: 'SLACK',
-  });
-  const [slack, setSlack] = useState<boolean>(true);
-  const [email, setEmail] = useState<boolean>(false);
-  const { racketType, statusMessage, snsNotiOpt } = profile;
-
-  useEffect(() => {
-    setEditedProfile((prev) => ({
-      ...prev,
-      racketType,
-      statusMessage,
-      snsNotiOpt,
-    }));
-    if (snsNotiOpt === 'BOTH') {
-      setSlack(true);
-      setEmail(true);
-    } else if (snsNotiOpt === 'SLACK') {
-      setSlack(true);
-      setEmail(false);
-    } else if (snsNotiOpt === 'EMAIL') {
-      setSlack(false);
-      setEmail(true);
-    } else {
-      setSlack(false);
-      setEmail(false);
-    }
-  }, [profile]);
-
-  const inputChangeHandler = ({
-    target: { name, value },
-  }:
-    | React.ChangeEvent<HTMLTextAreaElement>
-    | React.ChangeEvent<HTMLInputElement>) => {
-    if (name === 'statusMessage' && value.length > CHAR_LIMIT) {
-      value = value.slice(0, CHAR_LIMIT);
-      setEditedProfile((prev) => ({
-        ...prev,
-        [name]: value,
-      }));
-    }
-  };
 
   const finishEditHandler = async () => {
     if (slack && email) {
@@ -103,15 +50,7 @@ const useEditProfileModal = (): useEditProfileModalResult => {
 
   const cancelEditHandler = () => setModal({ modalName: null });
 
-  return [
-    slack,
-    email,
-    setEmail,
-    editedProfile,
-    inputChangeHandler,
-    finishEditHandler,
-    cancelEditHandler,
-  ];
+  return [finishEditHandler, cancelEditHandler];
 };
 
 export default useEditProfileModal;
